requirements.txt
flask
pymongo[srv]
gunicorn
python-dotenv